+++ jgclark.DashboardReact/src/react/components/ItemGrid.jsx
@@ -6,7 +6,6 @@
 
 import React, { useState } from 'react'
 import type { TSectionItem, TSection } from '../../types.js'
-import useInteractiveProcessing from '../customHooks/useInteractiveProcessing.jsx'
 import ItemRow from './ItemRow.jsx'
 import { useAppContext } from './AppContext.jsx'
 import { logDebug } from '@helpers/react/reactDev.js'
@@ -18,9 +17,6 @@ type Props = {
 
 function ItemGrid({ items, thisSection }: Props): React$Node {
   const { reactSettings, setReactSettings, sendActionToPlugin } = useAppContext()
-  const [itemsCopy, setItemsCopy] = useState<Array<TSectionItem>>([])
-
-  useInteractiveProcessing(items, thisSection, itemsCopy, setItemsCopy, reactSettings, setReactSettings, sendActionToPlugin)
 
   const visibleItems = items.map((item) => <ItemRow key={item.ID} item={item} thisSection={thisSection} />)